// packages/components/src/components/row/row.types.ts
export enum RowGutter {
    None = 'none',
    Xs = 'xs',
    Sm = 'sm',
    Md = 'md',
    Lg = 'lg',
    Xl = 'xl',
}

export enum RowHorizontalAlign {
    Around = 'around',
    Between = 'between',
    Center = 'center',
    Evenly = 'evenly',
    Left = 'left',
    Right = 'right',
}

export enum RowVerticalAlign {
    Bottom = 'bottom',
    Middle = 'middle',
    Stretch = 'stretch',
    Top = 'top',
}

export enum RowCssClasses {
    Base = 'example-row',
    HorizontalAlign = 'example-row__h-align',
    VerticalAlign = 'example-row__v-align',
    Gutter = 'example-row__gutter',
}